chuyen logout qua trai , chinh html css, nhan biet user dang duoc chon
