+++ src/components/leftSideBar/ChatList.jsx
@@ -1,7 +1,8 @@
-import React, {useEffect, useState} from 'react';
+import React, { useEffect, useState } from 'react';
 import '../../assets/style/chatList.css';
 
 const formatDateTime = (dateTimeString) => {
+    if (!dateTimeString) return ''; // Handling edge case where dateTimeString is null or undefined
     const dateTime = new Date(dateTimeString);
     const dateOptions = { year: 'numeric', month: 'numeric', day: 'numeric' };
     const timeOptions = { hour: 'numeric', minute: 'numeric', second: 'numeric' };
@@ -12,43 +13,40 @@ const formatDateTime = (dateTimeString) => {
     return `${datePart} ${timePart}`;
 };
 
-const ChatList = ({ users,handleUserClick, selectedUser, searchQuery }) => {
-
-    const [showChat, setShowChat] = useState(false);
+const ChatList = ({ users, handleUserClick, selectedUser, searchQuery }) => {
     const [filteredUserList, setFilteredUserList] = useState([]);
+    const [selectedItem, setSelectedItem] = useState(null); // State to track selected user/room
 
     useEffect(() => {
-        if (searchQuery !== undefined && searchQuery !== "") {
-            const filteredList = users.filter(
-                (user) =>
-                    user.name !== undefined &&
-                    user.name.toLowerCase().includes(searchQuery.toLowerCase())
+        if (searchQuery && searchQuery.trim() !== "") {
+            const filteredList = users.filter(user =>
+                user.name && user.name.toLowerCase().includes(searchQuery.trim().toLowerCase())
             );
             setFilteredUserList(filteredList);
         } else {
             setFilteredUserList(users);
         }
     }, [users, searchQuery]);
-    function handleUserItemClick(userName, userType) {
+
+    const handleUserItemClick = (userName, userType) => {
         handleUserClick(userName, userType);
-        setShowChat(true);
-    }
+        setSelectedItem(userName); // Set the selected user (you can modify this for rooms as well)
+    };
 
     return (
         <ul>
-            {users.map((user, index) => (
-                <li key={index} className="items"
+            {filteredUserList.map((user, index) => (
+                <li key={index} className={`items ${user.name === selectedItem ? 'selected' : ''}`}
                     onClick={() => handleUserItemClick(user.name, user.type)}>
 
-                    <img src={user.type===0 ? "/img/avata.png" : "/img/avatamuti.png"}
-                         alt="" />
+                    <img src={user.type === 0 ? "/img/avata.png" : "/img/avatamuti.png"} alt="" />
                     <div className="texts">
                         <span>{user.name}</span>
                         <p>{formatDateTime(user.actionTime)}</p>
                     </div>
                 </li>
             ))}
-            {users.length === 0 && <p className="no-users">Chưa có danh sách liên hệ!</p>}
+            {filteredUserList.length === 0 && <p className="no-users">Chưa có danh sách liên hệ!</p>}
         </ul>
     );
 };
